webadmin/fitcrackAPI/src/requirements.txt
flask-restplus
Flask-SQLAlchemy
pymysql
flask-cors
flask-login
flask-migrate
flask_script